// src/enum/enum.ts
export enum Filter {
   Count = 'count',
   Year = 'year'
}

export enum Filters {
   Ball = 'шар',
   Bell = 'колокольчик',
   Cone = 'шишка',
   Snow = 'снежинка',
   Toy = 'фигурка',
   White = 'белый',
   Yellow = 'желтый',
   Red = 'красный',
   Blue = 'синий',
   Green = 'зелёный',
   Big = 'большой',
   Average = 'средний',
   Small = 'малый'
}

export enum FiltersType {
   MultiSelect = "multiSelect",
   Range = "range"
}

export enum Sort {
   SortQueue = 'sort-queue',
   SortNameMax = 'sort-name-max',
   SortNameMin = 'sort-name-min',
   SortCountMax = 'sort-count-max',
   SortCountMin = 'sort-count-min'
}